Update fastJsonParser.c
Resolved a bug because of which for two digit PRIMITIVE the End pointer of fast_json_parser would point to the first digit. This is solved now. 
Also resolved a bug where the PRIMITIVE in a "}]" or more deep would stop parsing and it would stuck at JSON_UNDEFINED. For it added a new case for ",". So this is resolved now but some iterations/calls to json_parse would return JSON_UNDEFINED which needs to be skipped in application side.